rename: 트리 등록 하위 페이지명 의미에 맞는 이름으로 수정
