+++ src/pages/TreeRegistPage/LocationPickerMap/components/Map/style.ts
