+++ src/pages/TreeRegistPage/LocationPickerMap/style.ts
