+++ src/pages/TreeRegistPage/TreeDetailForm/components/LocationDetail/style.ts
