+++ src/pages/TreeRegistPage/TreeDetailForm/components/common/ButtonsInput/style.ts
