+++ src/pages/TreeRegistPage/TreeDetailForm/components/common/style.ts
