+++ src/pages/TreeRegistPage/TreeDetailForm/components/style.ts
