+++ src/routes/Router.tsx
@@ -6,11 +6,11 @@ import { MyPage } from 'pages/MyPage';
 import { LandingPage } from 'pages/LandingPage';
 import SignIn from 'pages/LoginPage/SocialLogin';
 import PATH from 'constants/path';
-import SearchLocation from 'pages/TreeRegistPage/Search';
+import SearchLocation from 'pages/TreeRegistPage/AddressSearch';
 import Redirect from 'pages/LoginPage/Redirect';
 import ReviewDetailPage from 'pages/ReviewDetailPage';
-import RegistMap from 'pages/TreeRegistPage/Map';
-import TreeRegiDetail from 'pages/TreeRegistPage/Form';
+import RegistMap from 'pages/TreeRegistPage/LocationPickerMap';
+import TreeRegiDetail from 'pages/TreeRegistPage/TreeDetailForm';
 import Nickname from 'pages/LoginPage/ProfileSetting';
 import SavePage from 'pages/SavedTreePage';
 import RegistedTreePage from 'pages/MyTreePage';
